Add `Reminder` to disable battery optimisation
Implement the `Reminder` subclass `BatteryOptimisationReminder`
that, in the event a device is running Android Marshmallow or above
and lacks Google's Play Services, informs the user they can disable
battery optimisation for Signal.

Recent releases of Android begin to restrict applications over time
if the device they are on is not moved, or interacted with for an
extend period; this includes access to network communication. When
this happens devices without access to Google Cloud Messaging won't
receive messages.

Signed-off-by: Alex Melbourne <alex.melbourne@protonmail.com>